python3_tutorial/07_api_basics/requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
folium>=0.14.0
pandas>=2.0.0
typer>=0.9.0
rich>=13.0.0
redis>=4.6.0
asyncio-mqtt>=0.13.0
schedule>=1.2.0
python-dateutil>=2.8.0